deepwish/requirements.txt
torch>=2.3.0
triton>=3.0.0
transformers>=4.44.0

pandas>=2.1.0
numpy>=1.24.0
pyyaml>=6.0.0
matplotlib>=3.7.0

tensorboard>=2.15.0
pytest>=7.4.0

sentencepiece>=0.1.99
safetensors>=0.4.0
